zmieniono: umiejscowiono interfejsu w lokalizacji plików z klasami
