+++ CkpTodoApp/DatabaseControllers/IDatabaseManagerInterface.cs
@@ -1,0 +1,15 @@
+﻿namespace CkpTodoApp.DatabaseControllers
+{
+    public interface IDatabaseManagerInterface
+    {
+        void InitDatabase();
+
+        void SeedDatabase();
+
+        void ExecuteSQL(string sql);
+
+        string ExecuteSQLQuery(string sql);
+
+        string DatabasePath();
+    }
+}
