+++ CkpTodoApp/DatabaseControllers/ISeederInterface.cs
@@ -1,0 +1,9 @@
+﻿namespace CkpTodoApp.DatabaseControllers
+{
+    public interface ISeederInterface
+    {
+        void MigrateDatabase();
+
+        void SeedDatabase();
+    }
+}
