--- CkpTodoApp/Interfaces/ISeederInterface.cs
@@ -1,9 +1,0 @@
-﻿namespace CkpTodoApp.Interfaces
-{
-  public interface ISeederInterface
-  {
-    void MigrateDatabase();
-
-    void SeedDatabase();
-  }
-}
