+++ CkpTodoApp/Models/IApiUserInterface.cs
@@ -1,0 +1,15 @@
+﻿namespace CkpTodoApp.Models
+{
+    public interface IApiUserInterface
+    {
+        int Id { get; set; }
+
+        string Name { get; set; }
+
+        string Surname { get; set; }
+
+        string Email { get; set; }
+
+        string PasswordHashed { get; set; }
+    }
+}
